wine-app/api/requirements.txt
# Flask and extensions
flask>=3.0.0
flask-cors>=4.0.0
flask-jwt-extended>=4.6.0

# OAuth
authlib>=1.3.0
httpx>=0.27.0

# Database
sqlalchemy>=2.0.0
alembic>=1.13.0
psycopg2-binary>=2.9.0

# Data validation
pydantic>=2.0.0

# AI/ML
openai>=1.0.0
pinecone-client>=3.0.0

# Utilities
python-dotenv>=1.0.0
